#5: Replace Python with Janino
Added an UpdateSystem to trigger generic component updates

+++ eris-scene/src/main/java/org/homonoia/eris/ecs/Component.java
@@ -12,5 +12,5 @@ public abstract class Component {
 
     private Entity entity;
 
-    public void update(float delta) {}
+    public void update(double delta) {}
 }
